ref oneis/haplo#6, use pg_config.
This provides an improvement over hard coding paths, as in the
worse case where pg_config can't be found, the path can be
modified before running the script (the other option is being
forced to edit the script itself).

pg_config should already be on the path for most working postgres
installs. (found this when using the popular Postgres.app on OS X)